refactor: move optional companion env-secret resume refresh into companion_env sibling
Relocate _refresh_optional_companion_env_secrets_for_resume and its two
private helpers (_missing/_present_optional_companion_env_secret_refs) from
monitor_handoff.py into the monitor_handoff_companion_env sibling that already
owns the compose-payload primitives they call. monitor_handoff.py re-exports
them (matching the existing _companion_env alias block) so resume_pr_monitor's
bare-name calls and the test module-attribute access keep working unchanged.

This frees headroom under the 1,500-line first-party file gate
(monitor_handoff.py 1495 -> 1384) so the follow-up forge-reason-code fix can
add a dedicated except clause without crossing the limit. No behavior change;
the two tests that monkeypatch the moved seams (_atomic_write_text,
optional_env_secret_compose_placeholder) now patch the companion_env module
where the code lives, with assertions unchanged.
