+++ src/awf/control/executor/monitor_handoff_companion_env.py
@@ -13,6 +13,12 @@ import yaml
 from yaml.constructor import ConstructorError
 from yaml.nodes import MappingNode, ScalarNode
 
+from awf.control.executor.quality_gates import _log
+from awf.node.companion_services import (
+    WorkspaceCompanionSpec,
+    optional_env_secret_compose_placeholder,
+)
+
 
 class _ComposeInterpolationPreservingDumper(yaml.SafeDumper):
     """Safe YAML dumper that keeps Compose interpolation scalars active."""
@@ -224,3 +230,120 @@ def _compose_environment_list_item_targets(item: object, targets: set[str]) -> b
         return False
     key = item.split("=", 1)[0]
     return key in targets
+
+
+def _refresh_optional_companion_env_secrets_for_resume(
+    *,
+    workspace_id: str,
+    compose_file: Path,
+    companion_specs: tuple[WorkspaceCompanionSpec, ...],
+    environ: Mapping[str, str],
+) -> None:
+    """Refresh optional companion env-secret targets before compose resume."""
+    missing_targets = _missing_optional_companion_env_secret_targets(
+        companion_specs=companion_specs,
+        environ=environ,
+    )
+    present_refs = _present_optional_companion_env_secret_refs(
+        companion_specs=companion_specs,
+        environ=environ,
+    )
+    if not missing_targets and not present_refs:
+        return
+
+    try:
+        payload = _safe_load_compose_payload_for_resume(compose_file.read_text(encoding="utf-8"))
+    except OSError:
+        _log.warning(
+            "executor.resume_companion_env_secret_refresh_read_failed",
+            workspace_id=workspace_id,
+            compose_file=str(compose_file),
+        )
+        return
+    except yaml.YAMLError:
+        _log.warning(
+            "executor.resume_companion_env_secret_refresh_parse_failed",
+            workspace_id=workspace_id,
+            compose_file=str(compose_file),
+        )
+        return
+
+    # This best-effort resume repair uses PyYAML, which preserves Compose
+    # interpolation via the custom dumper but not comments or block-scalar style.
+    removed_count = _remove_compose_environment_targets(payload, missing_targets)
+    restored_count = _restore_compose_environment_refs(payload, present_refs)
+    if removed_count == 0 and restored_count == 0:
+        return
+
+    try:
+        _atomic_write_text(
+            compose_file,
+            _safe_dump_compose_payload_for_resume(payload),
+            encoding="utf-8",
+        )
+    except OSError:
+        _log.warning(
+            "executor.resume_companion_env_secret_refresh_write_failed",
+            workspace_id=workspace_id,
+            compose_file=str(compose_file),
+        )
+        return
+    _log.warning(
+        "executor.resume_companion_env_secret_refresh_reformatted",
+        workspace_id=workspace_id,
+        compose_file=str(compose_file),
+        removed_count=removed_count,
+        restored_count=restored_count,
+        detail=(
+            "optional companion env-secret refresh rewrote the compose file; "
+            "comments, block-scalar style, and explicit null markers are not preserved"
+        ),
+    )
+    if removed_count:
+        _log.info(
+            "executor.resume_companion_optional_env_secrets_omitted",
+            workspace_id=workspace_id,
+            compose_file=str(compose_file),
+            omitted_count=removed_count,
+        )
+    if restored_count:
+        _log.info(
+            "executor.resume_companion_optional_env_secrets_restored",
+            workspace_id=workspace_id,
+            compose_file=str(compose_file),
+            restored_count=restored_count,
+        )
+
+
+def _missing_optional_companion_env_secret_targets(
+    *,
+    companion_specs: tuple[WorkspaceCompanionSpec, ...],
+    environ: Mapping[str, str],
+) -> dict[str, set[str]]:
+    missing_targets: dict[str, set[str]] = {}
+    for spec in companion_specs:
+        for secret in spec.environment_secrets:
+            if secret.required or secret.provider != "env" or secret.kind != "env":
+                continue
+            if secret.value_from in environ:
+                continue
+            missing_targets.setdefault(spec.name, set()).add(secret.target)
+    return missing_targets
+
+
+def _present_optional_companion_env_secret_refs(
+    *,
+    companion_specs: tuple[WorkspaceCompanionSpec, ...],
+    environ: Mapping[str, str],
+) -> dict[str, dict[str, str]]:
+    present_refs: dict[str, dict[str, str]] = {}
+    for spec in companion_specs:
+        for secret in spec.environment_secrets:
+            if secret.required or secret.provider != "env" or secret.kind != "env":
+                continue
+            if secret.value_from not in environ:
+                continue
+            present_refs.setdefault(spec.name, {})[secret.target] = (
+                optional_env_secret_compose_placeholder(secret.value_from)
+            )
+    return present_refs
